Portal: Make `PickFolder()` check that portal interface version is >=3 (#94)
This makes `NFD_PickFolder()` check that the xdg-desktop-portal interface version is at least 3 before requesting for a folder picker.  On earlier interface versions, it will return with an error.

This is because the `directory` flag on the `OpenFile` method of the portal was added in version 3.  On earlier versions, this flag will be ignored, which will cause a file picker (instead of folder picker) to be opened instead, and this is unlikely to be an appropriate fallback.