GTK: Open the dialog window on the same GdkScreen as the parent (#145)
Setting the default GdkDisplay doesn't actually cause a newly opened dialog window to use it. The more correct way seems to be to make the dialog window use the same GdkScreen as the parent. Each GdkScreen maps to one GdkDisplay, so this guarantees that the dialog window uses the same GdkDisplay as the parent too.